+++ CHANGELOG.md
@@ -37,6 +37,7 @@
 - Extend `wp.expect_near()` to support all vectors and quaternions.
 - Extend `wp.quat_from_matrix()` to support 4x4 matrices.
 - Updated the `OgnClothSimulate` node to use the VBD integrator ([GH-512](https://github.com/NVIDIA/warp/issues/512)).
+- Removed the `globalScale` parameter from the `OgnClothSimulate` node.
 
 ### Fixed